MEDIA/requirements.txt
Pillow
aiobotocore
aio_pika
pika